refactor: give Supabase requests more descriptive names
