+++ src/lib/requests/getRowsByTopic.ts
@@ -1,6 +1,6 @@
 import { DistrictLabel } from '@data/districts'
 import { supabase } from '@lib/requests/createSupabaseClient'
-import { HaushaltsdatenRowType } from './getMainTopicData'
+import { HaushaltsdatenRowType } from './getRowsByDistrictAndType'
 
 export type TopicColumnName =
   | 'hauptfunktions_bezeichnung'
@@ -8,31 +8,30 @@ export type TopicColumnName =
   | 'funktions_bezeichnung'
 
 export interface GetRowsByTopicParamsType {
-  bereich?: DistrictLabel
-  titelart: 'Einnahmetitel' | 'Ausgabetitel'
+  district?: DistrictLabel
+  expenseType: 'Einnahmetitel' | 'Ausgabetitel'
   topicColumn?: TopicColumnName
   topicValue?: string
 }
 
 /**
- * Retrieves rows from the Haushaltdaten based on the provided topicKey and topicValue.
- * topicKey may be one of hauptfunktions_bezeichnung, oberfunktions_bezeichnung, funktions_bezeichnung.
+ * Retrieves rows from the Haushaltdaten based on the provided topicKey and topicValue. `topicKey` may be one of `hauptfunktions_bezeichnung`, `oberfunktions_bezeichnung`, and `funktions_bezeichnung`.
  */
 export const getRowsByTopic = async ({
-  bereich,
-  titelart,
+  district,
+  expenseType,
   topicColumn,
   topicValue,
 }: GetRowsByTopicParamsType): Promise<HaushaltsdatenRowType[] | undefined> => {
-  if (!!bereich && !!topicColumn && !!topicValue) {
+  if (!!district && !!topicColumn && !!topicValue) {
     const { data, error } = await supabase
       .from('haushaltsdaten_2022')
       .select(
         'id, betrag, bereichs_bezeichnung, titel_art, titel_bezeichnung, hauptfunktions_bezeichnung, oberfunktions_bezeichnung, funktions_bezeichnung'
       )
       .eq('jahr', '2022')
-      .eq('titel_art', titelart)
-      .eq('bereichs_bezeichnung', bereich)
+      .eq('titel_art', expenseType)
+      .eq('bereichs_bezeichnung', district)
       .eq(topicColumn, topicValue)
 
     if (error) throw error
@@ -45,21 +44,21 @@ export const getRowsByTopic = async ({
         'id, betrag, bereichs_bezeichnung, titel_art, titel_bezeichnung, hauptfunktions_bezeichnung, oberfunktions_bezeichnung, funktions_bezeichnung'
       )
       .eq('jahr', '2022')
-      .eq('titel_art', titelart)
+      .eq('titel_art', expenseType)
       .eq(topicColumn, topicValue)
 
     if (error) throw error
 
     return data as HaushaltsdatenRowType[]
-  } else if (bereich && !topicColumn && !topicValue) {
+  } else if (district && !topicColumn && !topicValue) {
     const { data, error } = await supabase
       .from('haushaltsdaten_2022')
       .select(
         'id, betrag, bereichs_bezeichnung, titel_art, titel_bezeichnung, hauptfunktions_bezeichnung, oberfunktions_bezeichnung, funktions_bezeichnung'
       )
       .eq('jahr', '2022')
-      .eq('titel_art', titelart)
-      .eq('bereichs_bezeichnung', bereich)
+      .eq('titel_art', expenseType)
+      .eq('bereichs_bezeichnung', district)
 
     if (error) throw error
 
@@ -71,7 +70,7 @@ export const getRowsByTopic = async ({
         'id, betrag, bereichs_bezeichnung, titel_art, titel_bezeichnung, hauptfunktions_bezeichnung, oberfunktions_bezeichnung, funktions_bezeichnung'
       )
       .eq('jahr', '2022')
-      .eq('titel_art', titelart)
+      .eq('titel_art', expenseType)
 
     if (error) throw error
